requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
pillow
boto3
python-dotenv
